+++ FinalProject/src/main/java/com/group1project/controller/OrderController.java
@@ -2,7 +2,6 @@ package com.group1project.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
 
 @Controller
 public class OrderController {
@@ -10,10 +9,5 @@ public class OrderController {
 	@GetMapping("cart")
 	public String cart() {
 		return "cart";
-	}
-	
-	@GetMapping("mycart")
-	public String mycart() {
-		return "mycart";
 	}
 }
