+++ FinalProject/src/main/java/com/group1project/controller/PageController.java
@@ -39,13 +39,13 @@ public class PageController {
 		return "addProduct";
 	}
 	
-	@GetMapping("/login/insert")
-	public String insertAccountPage(Model model) {
-		Account account = new Account();
-		
-		model.addAttribute("account", account);
-		return "addAccount";
-	}
+//	@GetMapping("/login/insert")
+//	public String insertAccountPage(Model model) {
+//		Account account = new Account();
+//		
+//		model.addAttribute("account", account);
+//		return "addAccount";
+//	}
 	
 //	@GetMapping("back/allProduct2")
 //	public String productAll() {
