+++ FinalProject/src/main/java/com/group1project/model/repository/AccountRepository.java
@@ -1,9 +1,10 @@
 package com.group1project.model.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import com.group1project.model.bean.Account;
-
+@Repository
 public interface AccountRepository extends JpaRepository<Account, Integer>{
