+++ FinalProject/src/main/java/com/group1project/model/service/impl/AccountServiceimpl.java
@@ -3,12 +3,17 @@ package com.group1project.model.service.impl;
 import java.util.List;
 import java.util.Optional;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
 import com.group1project.model.bean.Account;
 import com.group1project.model.repository.AccountRepository;
 import com.group1project.model.service.AccountService;
 
+@Service
 public class AccountServiceimpl implements AccountService {
 
+	@Autowired
 	private AccountRepository aDao;
 
 	@Override
